requirements.txt
Flask==3.1.2
flask-cors==6.0.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.10
python-dotenv==1.1.1
Werkzeug==3.1.3
flask-jwt-extended==4.7.1
flask-marshmallow==1.3.0